+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added live lookup from Radarr/Sonarr if missing cached \*arr data (sync hasn't been ran since they was added)
 - Don't scan library path after deletion if using media server to delete media directly
 - Scan erroring out when invalid TMDB is parsed from a media server
+- Fixed failure when migrating DB (table _alembic_tmp_notification_settings already exists)
 
 ### Removed